Translate: move split_cmd_msg to PSO2DataTools Also handle messages with no spaces
